remove assets module setup
